==> scripts/1_compile_blink_data.Rmd <==
---
title: "Blink Data Compilation"
author: "Megan Schumer"
date: '`r Sys.Date()`'
output: pdf_document
toc: TRUE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(tibble)
```

```{r compile R56 pupil cloud blinks}

# list blink files
file_paths_r56 <- list.files("../data/raw/R56", pattern = "blinks.csv", full.names = TRUE, recursive = TRUE)

# create empty df to hold blinks
pupil_r56 <- tibble()

# loop through files and process
for (file_path in file_paths_r56) {
  
  # get rid of file path
  file_name_parts <- unlist(strsplit(file_path, split = "/|-"))
  id <- file_name_parts[5]
  participant <- file_name_parts[6]

  # Check if the file exists before attempting to read
  if (file.exists(file_path)) {
    
    # read the blinks.csv file
    data <- read_csv(file_path, col_types = cols())

    # Add the ID and pt columns
    data$id <- as.numeric(id)
    data$participant <- sub("_.*", "", participant)  # Remove everything after the underscore

    # Merge this file's data into the main dataframe
    pupil_r56 <- bind_rows(pupil_r56, data)
  }
}

# clean names
pupil_r56 <- janitor::clean_names(pupil_r56)

# repeat above steps to generate TSST timestamp df

# list event files
file_paths_r56_events <- list.files("../data/raw/R56", pattern = "events.csv", full.names = TRUE, recursive = TRUE)

# create empty df to hold timestamps
pupil_r56_timestamps <- tibble()

# loop through files and process
for (file_path in file_paths_r56_events) {
  
  # add ID and pt columns from folder names
  file_name_parts <- unlist(strsplit(file_path, split = "/|-"))
  id <- file_name_parts[5]
  participant <- file_name_parts[6]
  
  # check if file exists before attempting to read
  if (file.exists(file_path)) {
    
    # read the events.csv file
    data <- read_csv(file_path, col_types = cols())
    
    # add the ID and pt columns
    data$id <- as.numeric(id)
    data$participant <- sub("_.*", "", participant)  # Remove everything after the underscore
    
    # merge into main df
    pupil_r56_timestamps <- bind_rows(pupil_r56_timestamps, data)
  }
}

# remove event type columns
pupil_r56_timestamps <- select(pupil_r56_timestamps, 1:3, 5:6) %>%
  
  # pivot the timestamps wider, one row per pt
  pivot_wider(names_from = name, values_from = "timestamp [ns]") %>%
  
  # clean names
  janitor::clean_names()
```

```{r reformat r56 pupil blinks}

# add timestamps to pupil_r56_export
pupil_r56_ts <- left_join(pupil_r56, pupil_r56_timestamps, join_by(id == id, participant == participant, recording_id == recording_id)) %>%
  
  # categorize blinks into different task sections
  mutate(blink_type = case_when(
    start_timestamp_ns >= pprep_start & end_timestamp_ns <= pprep_end ~ "pprep",
    start_timestamp_ns >= pspeech_start & end_timestamp_ns <= pspeech_end ~ "pspeech",
    start_timestamp_ns >= cprep_start & end_timestamp_ns <= cprep_end ~ "cprep",
    start_timestamp_ns >= cspeech_start & end_timestamp_ns <= cspeech_end ~ "cspeech",
    .default = NA
  )) %>%
  
  # remove blinks outside of task sections
  filter(!is.na(blink_type)) %>%
  
  # convert UTC to relative timestamp in s
  mutate(start_timestamp_s = (start_timestamp_ns - recording_begin)/(10^9), end_timestamp_s = (end_timestamp_ns - recording_begin)/(10^9)) %>%
  
  # select relevant columns
  select(id, participant, start_timestamp_s, end_timestamp_s, duration_ms, blink_type)
```

```{r compile R01 pupil cloud blinks}

# list blink files
file_paths_r01 <- list.files("../data/raw/R01", pattern = "blinks.csv", full.names = TRUE, recursive = TRUE)

# create empty df to hold blinks
pupil_r01 <- tibble()

# loop through files and process
for (file_path in file_paths_r01) {
  
  # get rid of file path
  file_name_parts <- unlist(strsplit(file_path, split = "/|_"))
  id <- file_name_parts[5]
  participant <- file_name_parts[6]

  # Check if the file exists before attempting to read
  if (file.exists(file_path)) {
    
    # read the blinks.csv file
    data <- read_csv(file_path, col_types = cols())

    # Add the ID and pt columns
    data$id <- as.numeric(id)
    data$participant <- sub("_.*", "", participant)  # Remove everything after the underscore

    # Merge this file's data into the main dataframe
    pupil_r01 <- bind_rows(pupil_r01, data)
  }
}

# clean names
pupil_r01 <- janitor::clean_names(pupil_r01)

# repeat above steps to generate TSST timestamp df

# list event files
file_paths_r01_events <- list.files("../data/raw/R01", pattern = "events.csv", full.names = TRUE, recursive = TRUE)

# create empty df to hold timestamps
pupil_r01_timestamps <- tibble()

# loop through files and process
for (file_path in file_paths_r01_events) {
  
  # add ID and pt columns from folder names
  file_name_parts <- unlist(strsplit(file_path, split = "/|_"))
  id <- file_name_parts[5]
  participant <- file_name_parts[6]
  
  # check if file exists before attempting to read
  if (file.exists(file_path)) {
    
    # read the events.csv file
    data <- read_csv(file_path, col_types = cols())
    
    # add the ID and pt columns
    data$id <- as.numeric(id)
    data$participant <- sub("_.*", "", participant)  # Remove everything after the underscore
    
    # merge into main df
    pupil_r01_timestamps <- bind_rows(pupil_r01_timestamps, data)
  }
}

# remove event type columns
pupil_r01_timestamps <- select(pupil_r01_timestamps, 1:3, 5:6) %>%
  
  # pivot the timestamps wider, one row per pt
  pivot_wider(names_from = name, values_from = "timestamp [ns]") %>%
  
  # clean names
  janitor::clean_names()
```

```{r reformat r01 pupil blinks}

# add timestamps to pupil_r01_export
pupil_r01_ts <- left_join(pupil_r01, pupil_r01_timestamps, join_by(id == id, participant == participant, recording_id == recording_id)) %>%
  
  # categorize blinks into different task sections
  mutate(blink_type = case_when(
    start_timestamp_ns >= pprep_start & end_timestamp_ns <= pprep_end ~ "pprep",
    start_timestamp_ns >= pspeech_start & end_timestamp_ns <= pspeech_end ~ "pspeech",
    start_timestamp_ns >= cprep_start & end_timestamp_ns <= cprep_end ~ "cprep",
    start_timestamp_ns >= cspeech_start & end_timestamp_ns <= cspeech_end ~ "cspeech",
    .default = NA
  )) %>%
  
  # remove blinks outside of task sections
  filter(!is.na(blink_type)) %>%
  
  # convert UTC to relative timestamp in s
  mutate(start_timestamp_s = (start_timestamp_ns - recording_begin)/(10^9), end_timestamp_s = (end_timestamp_ns - recording_begin)/(10^9)) %>%
  
  # select relevant columns
  select(id, participant, start_timestamp_s, end_timestamp_s, duration_ms, blink_type)
```

```{r all blinks}

# combine r01 and r56 blinks into one table
all_blinks <- bind_rows(pupil_r56_ts, pupil_r01_ts) %>%
  
  # group by id, pt, and blink type
  group_by(id, participant, blink_type) %>%
  
  # count blinks for each pt by task section
  summarise(blink_count = n()) %>%
  
  # pivot wider, one row per pt per task section
  pivot_wider(names_from = blink_type, values_from = blink_count) %>%
  
  # write to csv
  write_csv("../data/raw/complete_blink_data.csv")

# identify potential errors in timestamping
all_blinks_na <- filter(all_blinks, is.na(cprep) | is.na(cspeech) | is.na(pprep) | is.na(pspeech))
```
