==> scripts/3_summary_stats.Rmd <==
---
title: "Data Integration"
author: "Megan Schumer"
date: '`r Sys.Date()`'
output: pdf_document
toc: TRUE
---

## start-up

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## setup

```{r packages, include=FALSE}

# load libraries
library(tidyverse)
library(readxl)

# variable for tidy data file path
existing_folder <- "../data/tidy"

# identify date/folder name of most recent tidy data output

# list files
tidy_folders <- list.dirs("../data/tidy", full.names = FALSE)
# identify most recent date
current_data <- tail(tidy_folders, 1)

# pull dataset from most recent data output
data_path <- file.path(existing_folder, current_data)
data_complete <- read_csv(paste0(data_path, "/data_complete.csv"))
```

## summary stats

```{r group data}

# calculate average age
summary_data_ages <- select(data_complete, 3:5) %>%
  
  # pivot age columns longer for summarize function
  pivot_longer(cols = c(2:3), names_to = "age") %>%
  
  # group by parent/child and variable_name
  group_by(parent_child, age) %>%
  
  # calculate mean/SD for each variable
  summarize(mean = mean(value, na.rm = TRUE), stdev = sd(value, na.rm = TRUE), count = sum(!is.na(value)), missing_values = sum(is.na(value)))

# count parent and child sex
summary_data_sex <- select(data_complete, 3, 6) %>%
  group_by(parent_child, sex) %>%
  summarize(count = n())

# count dyad sexes
summary_data_dyad_sex <- select(data_complete, 1, 3, 6) %>%
  mutate(parent_child = case_when(
    parent_child == 1 ~ "child",
    parent_child == 0 ~ "parent"
  )) %>%
  pivot_wider(names_from = parent_child, values_from = sex) %>%
  mutate(combo = paste0(parent, "-", child)) %>%
  group_by(combo) %>%
  summarize(count = n())

# race frequencies
summary_data_race <- select(data_complete, 3, 7) %>%
  group_by(parent_child, race) %>%
  summarize(count = n())

# ethnicity frequencies
summary_data_ethnicity <- select(data_complete, 3, 8) %>%
  group_by(parent_child, ethnicity) %>%
  summarize(count = n())

# parent education
summary_data_education <- select(data_complete, 3, 9) %>%
  filter(parent_child == 0) %>%
  group_by(parent_edu) %>%
  summarize(count = n())

# parent income
summary_data_income <- select(data_complete, 3, 10) %>%
  filter(parent_child == 0) %>%
  group_by(parent_income) %>%
  summarize(count = n())
```
